优化脚本

+++ ibas/deploy_apps.sh
@@ -60,7 +60,7 @@ echo ----------------------------------------------------
 # 部署顺序排序
 if [ ! -e "${IBAS_PACKAGE}/ibas.deploy.order.txt" ]; then
   cd ${IBAS_PACKAGE}
-  ls -l *.war | awk '//{print $NF}' >>"${IBAS_PACKAGE}/ibas.deploy.order.txt"
+  ls -ltr *.war | awk '//{print $NF}' >>"${IBAS_PACKAGE}/ibas.deploy.order.txt"
   cd ${WORK_FOLDER}
 fi
 echo 开始解压模块，到目录${IBAS_DEPLOY}
@@ -75,38 +75,43 @@ while read file; do
   grep -q ${folder} "${IBAS_DEPLOY}/ibas.release.txt" || echo "${folder}" >>"${IBAS_DEPLOY}/ibas.release.txt"
   # 解压war包到目录
   unzip -o "${IBAS_PACKAGE}/${file}" -d "${IBAS_DEPLOY}/${folder}"
-  # 删除配置文件，并映射到统一位置
-  if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/app.xml" ]; then
-    if [ ! -e "${IBAS_CONF}/app.xml" ]; then cp -f "${IBAS_DEPLOY}/${folder}/WEB-INF/app.xml" "${IBAS_CONF}/app.xml"; fi
-    rm -f "${IBAS_DEPLOY}/${folder}/WEB-INF/app.xml"
-    ln -s "${IBAS_CONF}/app.xml" "${IBAS_DEPLOY}/${folder}/WEB-INF/app.xml"
-  fi
-  # 删除服务路由文件，并映射到统一位置
-  if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/service_routing.xml" ]; then
-    if [ ! -e "${IBAS_CONF}/service_routing.xml" ]; then cp -f "${IBAS_DEPLOY}/${folder}/WEB-INF/service_routing.xml" "${IBAS_CONF}/service_routing.xml"; fi
-    rm -f "${IBAS_DEPLOY}/${folder}/WEB-INF/service_routing.xml"
-    ln -s "${IBAS_CONF}/service_routing.xml" "${IBAS_DEPLOY}/${folder}/WEB-INF/service_routing.xml"
+  # 存在WEB安全目录
+  if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/" ]; then
+    # 删除配置文件，并映射到统一位置
+    if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/app.xml" ]; then
+      if [ ! -e "${IBAS_CONF}/app.xml" ]; then cp -f "${IBAS_DEPLOY}/${folder}/WEB-INF/app.xml" "${IBAS_CONF}/app.xml"; fi
+      rm -f "${IBAS_DEPLOY}/${folder}/WEB-INF/app.xml"
+      ln -s "${IBAS_CONF}/app.xml" "${IBAS_DEPLOY}/${folder}/WEB-INF/app.xml"
+    fi
+    # 删除服务路由文件，并映射到统一位置
+    if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/service_routing.xml" ]; then
+      if [ ! -e "${IBAS_CONF}/service_routing.xml" ]; then cp -f "${IBAS_DEPLOY}/${folder}/WEB-INF/service_routing.xml" "${IBAS_CONF}/service_routing.xml"; fi
+      rm -f "${IBAS_DEPLOY}/${folder}/WEB-INF/service_routing.xml"
+      ln -s "${IBAS_CONF}/service_routing.xml" "${IBAS_DEPLOY}/${folder}/WEB-INF/service_routing.xml"
+    fi
+    # 映射日志文件夹到统一位置
+    if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/logs" ]; then rm -rf "${IBAS_DEPLOY}/${folder}/WEB-INF/logs"; fi
+    ln -s "${IBAS_LOG}" "${IBAS_DEPLOY}/${folder}/WEB-INF/"
+    # 映射数据文件夹到统一位置
+    if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/data" ]; then rm -rf "${IBAS_DEPLOY}/${folder}/WEB-INF/data"; fi
+    ln -s "${IBAS_DATA}" "${IBAS_DEPLOY}/${folder}/WEB-INF/"
+    # 集中共享jar包
+    if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/lib/" ]; then
+      if [ -e "${IBAS_LIB}" ]; then
+        # 记录文件名称
+        ls -l "${IBAS_DEPLOY}/${folder}/WEB-INF/lib/" >"${IBAS_DEPLOY}/${folder}/WEB-INF/lib/~file_list.txt"
+        # 复制模块jar包到tomcat的lib目录
+        cp -f "${IBAS_DEPLOY}/${folder}/WEB-INF/lib/"*.jar "${IBAS_LIB}"
+        # 清除tomcat的lib已经存在的jar包
+        rm -f "${IBAS_DEPLOY}/${folder}/WEB-INF/lib/"*.jar
+      fi
+    fi
   fi
   # 删除前端配置文件，并映射到统一位置
   if [ -e "${IBAS_DEPLOY}/${folder}/config.json" ]; then
     if [ ! -e "${IBAS_CONF}/config.json" ]; then cp -f "${IBAS_DEPLOY}/${folder}/config.json" "${IBAS_CONF}/config.json"; fi
     rm -f "${IBAS_DEPLOY}/${folder}/config.json"
     ln -s "${IBAS_CONF}/config.json" "${IBAS_DEPLOY}/${folder}/config.json"
-  fi
-  # 映射日志文件夹到统一位置
-  if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/logs" ]; then rm -rf "${IBAS_DEPLOY}/${folder}/WEB-INF/logs"; fi
-  ln -s "${IBAS_LOG}" "${IBAS_DEPLOY}/${folder}/WEB-INF/"
-  # 映射数据文件夹到统一位置
-  if [ -e "${IBAS_DEPLOY}/${folder}/WEB-INF/data" ]; then rm -rf "${IBAS_DEPLOY}/${folder}/WEB-INF/data"; fi
-  ln -s "${IBAS_DATA}" "${IBAS_DEPLOY}/${folder}/WEB-INF/"
-  # 集中共享jar包
-  if [ -e "${IBAS_LIB}" ]; then
-    # 记录文件名称
-    ls -l "${IBAS_DEPLOY}/${folder}/WEB-INF/lib/" >"${IBAS_DEPLOY}/${folder}/WEB-INF/lib/~file_list.txt"
-    # 复制模块jar包到tomcat的lib目录
-    cp -f "${IBAS_DEPLOY}/${folder}/WEB-INF/lib/"*.jar "${IBAS_LIB}"
-    # 清除tomcat的lib已经存在的jar包
-    rm -f "${IBAS_DEPLOY}/${folder}/WEB-INF/lib/"*.jar
   fi
   # 备份程序包
   mv "${IBAS_PACKAGE}/${file}" ${IBAS_PACKAGE_BACKUP}/${file}
@@ -124,7 +129,10 @@ fi
 # 共享jar包旧版清理
 if [ -e "${IBAS_LIB}" ]; then
   echo 清理[ibas_lib]旧版文件
-  if [ -e "${IBAS_LIB}/~file_types.txt" ]; then rm -rf "${IBAS_LIB}/~file_types.txt" && touch "${IBAS_LIB}/~file_types.txt"; fi
+  if [ -e "${IBAS_LIB}/~file_types.txt" ]; then
+    rm -rf "${IBAS_LIB}/~file_types.txt"
+  fi
+  touch "${IBAS_LIB}/~file_types.txt"
   for file in $(ls -r "${IBAS_LIB}" | awk '//{print $NF}'); do
     file_type=!${file%-*}-!
     grep -l ${file_type} "${IBAS_LIB}/~file_types.txt" >/dev/null
